fix issue #7
During the completing read of `M-x dmenu`, when a user uniquely matches an executable filename, but that filename occurs more than once on `$PATH` (`exec-path`), the completing read will show more than one entry with the same (completed) value.  This is because the candidate list `dmenu--cache-executable-files` doesn't have unique entries.  This fix ensures that the candidate list has unique entries.