Put back some original code
The commit prior to this assumed that the completing read function would sort the candidate list (given to it).  However, some completing read functions do NOT do this, and when this is the case for some users, they would like to see their particular user history at the top of the completion candidates displayed to them in the UI when `M-x dmenu` is initially invoked.  Hence, some of the original code that handled this is being restored by this commit.