Restore more original code
The `pushnew` call will not satisfy the user requirement of putting the most recent commands at the top of the user input history.  The original code (which removes the most recent user input from the history -- no matter how deep in the history it occurs(!) -- and pushes it on to the head of history) does (or rather, did) satisfy that requirement.  Hence, it goes back in.